requirements.txt
# debugpy
pydevd-odoo
http